feat: create payment api with vnpay
fix: replace name column of entity

+++ server/src/main/java/com/utc2/cntt/major_assignment/self_ordering_restaurant/dto/request/IngredientRequestDTO.java
@@ -6,8 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -17,7 +15,7 @@ public class IngredientRequestDTO {
     private String name;
     private String description;
     private String unit;
-    private BigDecimal costPerUnit;
+    private Long costPerUnit;
     private IngredientStatus status;
     private int minimumQuantity;
 }
